app/src/main/java/com/foodlab/foodReservation/item/service/ItemService.java
package com.foodlab.foodReservation.item.service;

import com.foodlab.foodReservation.item.dto.request.UpdateItemRequest;
import com.foodlab.foodReservation.item.entity.Item;
import com.foodlab.foodReservation.item.repository.ItemRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
@RequiredArgsConstructor
public class ItemService {

    private final ItemRepository itemRepository;

    @Transactional
    public void updateItem(Long itemId, UpdateItemRequest updateItemRequest) {
        Item item = itemRepository.findById(itemId).orElseThrow(
                () -> new IllegalArgumentException("존재하지 않는 메뉴입니다."));
        item.updateItem(updateItemRequest.getName(), updateItemRequest.getPrice());
    }
}